Relax required ruby version (#29)
* Relax required ruby version

A recent change restricted the required ruby version for this project to 3.0.3 or above. We have a few projects running on 2.7 and notably the design-system tests against both 2.7 and 3. Limiting the version to 3 prevents us from using the new version in these contexts.

The rubocop docs for TargetRubyVersion suggest that this respects the project value if this is unset:

> Otherwise, RuboCop will then check your project for a series of files where the version may be specified already. The files that will be looked for are .ruby-version, Gemfile.lock, and *.gemspec. If Gemspec file has an array for required_ruby_version, the lowest version will be used. If none of the files are found a default version value will be used.

https://docs.rubocop.org/rubocop/configuration.html#setting-the-target-ruby-version

As the ruby version is a project specific thing it makes sense to defer to the project rather than explicitly setting this in shared rules.

This change removes `TargetRubyVersion` and widens the `required_ruby_version` again so we can continue to use this config against 2.7 versions of ruby.

* update version and dependencies

* update readme

Co-authored-by: Daniel Lewis <github@mrdaniellewis.co.uk>