Config proposals (#33)
* Add TargetRubyVersion to internal rubocop config

Avoids misleading comment in gemspec and reflects minimally supported
rubocop version.

* Count arrays and hashes as one-line

Count arrays and hashes as one-line when calculating method and class
length. We have these set in many of our internal rubocop configs so it
makes sense to consider this a default preference.

* Add rubocop-rails as as dependency

As far as I know there's no reason this can't be added as a
dependency to avoid the extra step in the installation instructions.

* Bump dependency versions

* Remove ruby-version file

This is a shared gem that is intended to run on any version of ruby 2.7
and upwards. Including a ruby-version file seems counter to the
conventions of similar gems so perhaps we should remove this.

* Remove unneeded spec exclude

Remove `spec/` exclude for `Metrics/BlockLength` as this is now a default in rubocop-rspec 2.11+

* Use rubocop ~> 1.43

* update minimum to latest versions

---------

Co-authored-by: Daniel Lewis <github@mrdaniellewis.co.uk>