Add missing 'rubocop-capybara' dependency (#36)
With the recent v11.0.0 release, a line was added to default.yml to
disable the `Capybara/ClickLinkOrButtonStyle` cop. This resulted in a
`unrecognized cop or department Capybara/ClickLinkOrButtonStyle` error
to occur in projects which didn't happen to have a dependency on a new
enough version of rubocop-capybara (v2.19.0+).

Additionally, API-only Rails projects generally don't use Capybara in
their test suites, so the dependency was missing altogether.

To resolve this issue, I've added `rubocop-capybara` as an explicit
dependency to this gem and ensured that it is required in default.yml.